Don't show validator links on withdrawals page without beacon client
